Support release specific config file templates
This includes tools for building sample config files and updates the
existing designate and heat support to be able to find them in the new
release specific location.  This also adds new facts for the current
release name of the project and the current image name.  The release
name fact is used to ensure that the config file is replaced from the
template when upgrading/downgrading between releases.

Issue: CIRRUS-3177
Change-Id: Ib232c278f1fdf11a22159f8991a6dbade18a95a7