Add nova-compute class
Add new class for handling nova-compute in a container.  This has a new
before_start flag which is passed into the docker::run instantiation.
This can be used to ensure neutron-ovs-cleanup is run before
nova-compute is started.

This also changes the mode on nova directories to 755 instead of 750 so
that libvirt can access the /var/lib/nova directories.

Issue: CIRRUS-3877
Change-Id: I5d9a3c696c3b0ba5a558d2a58269091f9692c9db