Add bind mounts for iscsi in nova-compute container
The nova-compute process needs to discover what iscsi targets may be
available to it by examining the files in /etc/iscsi.  It will then
provide that information to cinder when requesting the volume attach if
found.  Once it has this information the iSCSI volume will be mounted
on the hypervisor and qemu will be told to mount the block device.  In
order for nova compute to validate this worked, it needs to be able to
access /dev/disk/by-path to see the block device.

Issue: CIRRUS-3967
Change-Id: Idea301bd8fdca14955f08bbb7775e9ae48eaa3d6