Relocate docker socket for 1.12
Support for not restarting containers on engine restart was added in
Docker 1.12.  This means that when providing access to the docker socket
to a container, that a simple bind mount is not enough, since that
socket will be stale after the docker engine is restarted.  Instead
mount a directory containing the docker socket into the container, and
then use the socket in that directory.  This change will also set
DOCKER_HOST inside the container to make dockercommands work seamlessly
inside the container.

Also updated the README with information about how to use the neutron
docker classes.

Change-Id: Ib55b91680f1e532a8c26cb58e6461c69e2023c8d