Add nova-compute support for NFS cinder backend
This introduces a new 'enable_nfs' flag that allows the cinder NFS
backend to work properly with nova compute inside a container.  Without
this the mounts are done inside the container, but not visible outside,
meaning that libvirt cannot see the volume to connect it to the
instance.  If the flag is enabled, then the nova::compute class will
make /var/lib/nova/mnt a bind mount so that the shared flag can be set
on this mount.  This allows us to tell docker to also set this up as a
shared mount inside of the container, so that the mounts will be
properly propagated.

Change-Id: I6ab5b90fbd4fa4666210f3b331bbdd253c0ac126