Fix ordering issues with Trove managed folders
When moving from packages to containers the ordering is currently as
follows:

1) This module checks and sets ownership on /var/log/trove
2) During the install phase the Trove packages are purged, which takes
/var/log/trove with it.
3) Later in the run other modules (specifically the dockerized db_sync)
try to use /var/log/trove, which now doesn't exist.

A 2nd puppet run solves this because the managed directories are
re-created, but we need to ensure this works in one run. Therefore this
change will create the managed folders after the install phase, meaning
after packages are purged, but before the db_sync phase when they are
needed.

Removing /var/log/trove seems to be a special behavior of Trove
packages, other OpenStack services have not seen this.

Change-Id: Ie7c60a8378ea0eec961354f827bc175435e4a8a7