Keystone SSL certificates and SSL vassal addition.
In order for Keystone to start SSL endpoints the docker
container will need to mount the SSL directory where
the proper certificates are stored.

We also need to configure the new vassals that will be
running the SSL endpoints. This will only take effect
when the /etc/keystone/keystone-paste.ini file is removed
on the keystone hosts and allowed to be regenerated with
puppet.

Issue: CIRRUS-5907
Change-Id: I42210a4edd9c300d2da15d959877791f6039166f